Document HorizonClient and Models
