--- src/horizon_client/mod.rs
@@ -1,1 +1,0 @@
-pub mod horizon_client;
