+++ src/lib.rs
@@ -1,103 +1,197 @@
-//! Stellar SDK for Rust
+//! Stellar Horizon SDK for Rust
 //!
-//! This is a Rust SDK for the Stellar Horizon API. It is a work in progress and is not yet ready for production use.
-//! This library defines traits for request and response objects. This library contains types that implement these traits for each endpoint in the Horizon API respectively.
-//! The request objects are used to build the request URL and the response objects are used to deserialize the response from JSON.
+//! This Rust library provides a user-friendly interface to the Stellar Horizon API, 
+//! allowing developers to easily query and transact on the Stellar network. Centered 
+//! around the `HorizonClient`, the SDK abstracts the underlying HTTP request and response 
+//! mechanisms into a set of simple, high-level methods.
 //!
-//! ## Examples
+//! The SDK is designed with a focus on developer experience, providing clear abstractions, 
+//! sensible defaults, and streamlined error handling.
 //!
-//! Here is an example on how to use this library:
+//! ## Status
+//!
+//! The SDK is under active development. It is functional but should be considered a 
+//! work-in-progress. Features may be added or changed, and the SDK may evolve before 
+//! stabilization.
+//! 
+//! #### Supported endpoints:
+//! ![25%](https://progress-bar.dev/25/?width=200)
+//! * Accounts
+//! * Assets
+//! * Claimable balance
+//! * Ledgers
+//! 
+//! #### Endpoints on the roadmap:
+//! * Effects
+//! * Fee stats
+//! * Liquidity pools
+//! * Offers
+//! * Operations
+//! * Orderbook
+//! * Paths
+//! * Payments
+//! * Trade aggregations
+//! * Trades
+//! * Transactions
+
+//!
+//! ## Example Usage
+//!
+//! The following example demonstrates how to use the `HorizonClient` to retrieve a list 
+//! of accounts with a specific signer:
 //!
 //! ```rust
+//! use stellar_rust_sdk::horizon_client::HorizonClient;
+//! use stellar_rust_sdk::accounts::prelude::{AccountsRequest, AccountsResponse};
+//! use stellar_rust_sdk::models::{Request, Response};
 //!
-//! use stellar_rust_sdk::horizon_client::horizon_client::HorizonClient;
-//! use stellar_rust_sdk::accounts::prelude::AccountsRequest;
-//! use stellar_rust_sdk::accounts::prelude::AccountsResponse;
-//! use crate::stellar_rust_sdk::models::Request;
+//! #[tokio::main]
+//! async fn main() -> Result<(), Box<dyn std::error::Error>> {
+//!     // Initialize the Horizon client with the testnet server
+//!     let horizon_client = HorizonClient::new("https://horizon-testnet.stellar.org".to_string())?;
 //!
-//! async fn example() -> Result<(), Box<dyn std::error::Error>> {
-//! // Initialize horizon client
-//! let horizon_client =
-//!     HorizonClient::new("https://horizon-testnet.stellar.org".to_string()).unwrap();
+//!     // Create a request to fetch accounts with a specific signer
+//!     let accounts_request = AccountsRequest::new()
+//!         .set_signer("GDQJUTQYK2MQX2VGDR2FYWLIYAQIEGXTQVTFEMGH2BEWFG4BRUY4CKI7")?
+//!         .set_limit(10)?;
 //!
-//! // construct request
-//! let mut accounts_request = AccountsRequest::new()
-//!     .set_signer("GDQJUTQYK2MQX2VGDR2FYWLIYAQIEGXTQVTFEMGH2BEWFG4BRUY4CKI7").unwrap()
-//!     .set_limit(10).unwrap();
+//!     // Perform the request using the Horizon client
+//!     let accounts_response: Result<AccountsResponse, String> =
+//!         horizon_client.get_account_list(&accounts_request).await;
 //!
-//! // call the get_account_list method to retrieve the account list response
-//! let _accounts_response: Result<AccountsResponse, String> =
-//!     horizon_client.get_account_list(&accounts_request).await;
+//!     // Check for success and handle the response or error
+//!     match accounts_response {
+//!         Ok(response) => {
+//!             // Process the response
+//!         },
+//!         Err(e) => {
+//!             // Handle the error
+//!         }
+//!     }
 //!
-//! assert!(_accounts_response.is_ok());
 //!     Ok(())
 //! }
 //! ```
-//! ## Features
 //!
-//! All the endpoints in the Horizon Api and wether or not they are supported by this library:
+//! This example initializes a `HorizonClient`, constructs an `AccountsRequest` to filter 
+//! accounts by signer, and calls `get_account_list` to retrieve the relevant data. 
+//! The result is then handled in a match expression, demonstrating the SDK's straightforward 
+//! error handling.
 //!
-//! - [x] `Accounts` - List all accounts endpoint
-//! - [x] `Single Account` - Get a single account endpoint
-//! - [x] `All Assets` - List all assets endpoint
-//! - [x] `All Claimable Balances` - List all claimable balances endpoint
-//! - [x] `Single Claimable Balance` - Get a single claimable balance endpoint
-//! - [ ] `All Effects` - List all effects endpoint
-//! - [ ] `Effects for Account` - List all effects for an account endpoint`
-//! - [ ] `Effects for Ledger` - List all effects for a ledger endpoint
-//! - [ ] `Effects for Operation` - List all effects for an operation endpoint
-//! - [ ] `Effects for Transaction` - List all effects for a transaction endpoint
-//! - [ ] `All Fee Stats` - List all fee stats endpoint
-//! - [x] `All Ledgers` - List all ledgers endpoint
-//! - [x] `Single Ledgers` - Get a single ledger endpoint
-//! - [ ] `All Liquidity Pools` - List all liquidity pools endpoint
-//! - [ ] `Single Liquidity Pool` - Get a single liquidity pool endpoint
-//! - [ ] `All Offers` - List all offers endpoint
-//! - [ ] `Single Offer` - Get a single offer endpoint
-//! - [ ] `Offers for Account` - List all offers for an account endpoint
-//! - [ ] `All Operations` - List all operations endpoint
-//! - [ ] `Single Operation` - Get a single operation endpoint
-//! - [ ] `Operations for Account` - List all operations for an account endpoint
-//! - [ ] `Operations for Ledger` - List all operations for a ledger endpoint
-//! - [ ] `Operations for Liquidity Pool` - List all operations for a liquidity pool endpoint
-//! - [ ] `Operations for Transaction` - List all operations for a transaction endpoint
-//! - [ ] `All Order Book Details` - List all order book endpoint
-//! - [ ] `Find payments Paths` - Find payment paths endpoint
-//! - [ ] `Find Strict Receive Payment Paths` - Find strict receive payment paths endpoint
-//! - [ ] `Find Strict Send Payment Paths` - Find strict send payment paths endpoint
-//! - [ ] `All Payments` - List all payments endpoint
-//! - [ ] `Payments for Account` - List all payments for an account endpoint
-//! - [ ] `Payments for Ledger` - List all payments for a ledger endpoint
-//! - [ ] `Payments for Transaction` - List all payments for a transaction endpoint
-//! - [ ] `Trade Aggregations` - List trade aggregations endpoint
-//! - [ ] `All Trades` - List all trades endpoint
-//! - [ ] `Trades for Account` - List all trades for an account endpoint
-//! - [ ] `Trades for Liquidity Pool` - List all trades for a liquidity pool endpoint
-//! - [ ] `Trades for Offer` - List all trades for an offer endpoint
-//! - [ ] `All Transactions` - List all transactions endpoint
-//! - [ ] `Single Transaction` - Get a single transaction endpoint
-//! - [ ] `Post Transaction` - Post transaction endpoint
-//! - [ ] `Transactions for Account` - List all transactions for an account endpoint
-//! - [ ] `Transactions for Ledger` - List all transactions for a ledger endpoint
-//! - [ ] `Transactions for Liquidity Pool` - List all transactions for a liquidity pool endpoint
+//! Visit the documentation for `HorizonClient` and endpoint-specific request and response 
+//! types for more examples and detailed usage instructions.
 
-/// The accounts module
+
+/// Requests and Response for retrieving accounts.
+///
+/// This module provides a set of specialized request and response structures designed for 
+/// interacting with the accounts-related endpoints of the Horizon server. These structures
+/// facilitate the construction of requests to query account data and the interpretation of
+/// the corresponding responses.
+/// 
+/// # Usage
+///
+/// This module is intended to be used in conjunction with the [`HorizonClient`](crate::horizon_client::HorizonClient) 
+/// for making specific account-related API calls to the Horizon server. The request 
+/// structures are designed to be passed to the client's methods, which handle the 
+/// communication with the server and return the corresponding response structures.
+///
+/// # Example
+/// An example of retrieving a list of accounts, filtering by signer:
+/// ```rust
+/// # use stellar_rust_sdk::accounts::prelude::{AccountsRequest, AccountsResponse};
+/// # use stellar_rust_sdk::models::Request;
+/// # use stellar_rust_sdk::horizon_client::HorizonClient;
+/// # 
+/// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
+/// # let base_url = "https://horizon-testnet.stellar.org".to_string();
+/// # let horizon_client = HorizonClient::new(base_url)
+/// #    .expect("Failed to create Horizon Client");
+/// let request = AccountsRequest::new()
+///     .set_signer("GDQJUTQYK2MQX2VGDR2FYWLIYAQIEGXTQVTFEMGH2BEWFG4BRUY4CKI7").unwrap()
+///     .set_limit(10).unwrap();
+/// 
+/// let response: Result<AccountsResponse, String> = horizon_client
+///     .get_account_list(&request)
+///     .await;
+/// # Ok({})
+/// # }
+/// 
 pub mod accounts;
-/// The assets module
-pub mod assets;
-/// The claimable balances module
-pub mod claimable_balances;
-/// The horizon client module
-pub mod horizon_client;
-/// The ledgers module
-pub mod ledgers;
-/// The models module, here the Request and Response traits are defined
-pub mod models;
-pub mod xdr;
 
-/// The asset type
-/// Native - The native asset
-/// Issued - An issued asset
+pub mod assets;
+
+pub mod claimable_balances;
+
+/// Client for calling the Stellar Horizon API 
+/// 
+/// # Constructing a `HorizonClient`
+/// A string containing the base URL for the Horizon API is required to contruct a client.
+/// For example, to construct a client for the Horizon API testnet:
+/// ```rust
+/// use stellar_rust_sdk::horizon_client::HorizonClient;
+/// 
+/// let base_url = "https://horizon-testnet.stellar.org".to_string();
+/// let horizon_client = HorizonClient::new(base_url)
+///     .expect("Failed to create Horizon Client");;
+/// ```
+/// 
+/// # Using the `HorizonClient`
+/// The HorizonClient has a function that can be called for each endpoind provided
+/// by the Horizon API. For example, it has a [`HorizonClient::get_account_list`](crate::horizon_client::HorizonClient::get_account_list) 
+/// function, which returns an async future that contains a result, as illustrated below:
+/// ```rust
+/// # use stellar_rust_sdk::assets::prelude::{AllAssetsRequest, AllAssetsResponse};
+/// # use stellar_rust_sdk::models::Request;
+/// # use stellar_rust_sdk::horizon_client::HorizonClient;
+/// # 
+/// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
+/// # let base_url = "https://horizon-testnet.stellar.org".to_string();
+/// # let horizon_client = HorizonClient::new(base_url)
+/// #    .expect("Failed to create Horizon Client");;
+/// let all_assets_request = AllAssetsRequest::new();
+/// let accounts_response: Result<AllAssetsResponse, String> = horizon_client
+///     .get_all_assets(&all_assets_request)
+///     .await;
+/// # Ok({})
+/// # }
+/// ```
+pub mod horizon_client;
+
+pub mod ledgers;
+
+/// Contains core data structures and traits.
+///
+/// This module is used by the Stellar Rust SDK to interact with the Horizon API. 
+/// It defines enums, traits, and functions that encapsulate the logic for 
+/// creating and processing HTTP requests and responses, as well as handling the 
+/// data involved in these operations.
+/// 
+/// The `models` module plays a critical role in abstracting the complexities 
+/// of the Horizon API, allowing developers to work with high-level Rust constructs 
+/// instead of raw HTTP requests and JSON responses.
+pub mod models;
+
+
+/// Represents the types of assets in the Stellar network.
+///
+/// `AssetType` is an enumeration used to specify the type of an asset in Stellar-related requests.
+/// It differentiates between native assets and issued assets within the Stellar ecosystem.
+///
+/// # Variants
+///
+/// * `Native` - Represents the native asset of the Stellar network (often referred to as XLM).
+/// * `Issued` - Represents an asset that is issued by an account on the Stellar network.
+///   In its current implementation, it does not hold the asset code and issuer account ID, 
+///   but future enhancements are intended to include these details for complete asset specification.
+///
+/// # Note
+///
+/// The `Issued` variant is currently a placeholder and does not encapsulate the complete 
+/// information required for an issued asset (i.e., Asset Code and Issuer Account ID). 
+/// This is a known limitation and should be addressed in future versions to ensure full 
+/// functionality.
+/// 
 #[derive(Clone)]
 pub enum AssetType {
     Native,
@@ -113,20 +207,52 @@ impl std::fmt::Display for AssetType {
     }
 }
 
+
+/// Extension trait for building query parameter strings from a vector of optional values.
+///
+/// This trait provides a method to construct a query string from a vector of optional 
+/// values (`Option<T>`). It is designed to be used for generating query parameters in 
+/// URL construction, where each parameter is only included if it has a value (`Some`).
+///
+/// # Usage
+/// This trait is typically used internally in constructing URLs with query parameters 
+/// by implementors of the [`Request::get_query_parameters`](crate::models::Request::get_query_parameters) 
+/// method. It enables a convenient and efficient way to handle optional parameters in 
+/// a URL query string.
+///
 trait BuildQueryParametersExt<T> {
+    /// Constructs a query string for an HTTP request from the object's properties.
+    ///
+    /// This method transforms the properties of the implementing object into a URL-encoded query 
+    /// string. Each property is converted to a key-value pair, and pairs are concatenated with '&'.
+    /// Properties that are `None` are omitted from the string.
+    ///
+    /// # Returns
+    /// A `String` representing the query parameters of the HTTP request. If there are no 
+    /// parameters, or all properties are `None`, an empty string is returned.
     fn build_query_parameters(self) -> String;
 }
 
 impl<T: ToString> BuildQueryParametersExt<Option<T>> for Vec<Option<T>> {
+    /// Constructs a query string from a vector of optional values.
     fn build_query_parameters(self) -> String {
-        let params = self
-            .into_iter()
-            // The filter_map function filters out the None values, leaving only the Some values with formatted strings.
-            .filter_map(|x| x.map(|val| val.to_string()))
+        let params = self.into_iter()
+            // Iterate over each element in the vector.
+            .filter_map(|x| 
+                // Use filter_map to process each Option<T>.
+                // If the element is Some, it's transformed to its string representation.
+                // If the element is None, it's filtered out.
+                x.map(|val| val.to_string()))
+            // Collect the transformed values into a Vec<String>.
             .collect::<Vec<String>>()
+            // Join the Vec<String> elements with '&' to create the query string.
             .join("&");
+
+        // Check if the resulting params string is empty.
         match params.is_empty() {
+            // If params is empty, return an empty string.
             true => "".to_string(),
+            // If params is not empty, prepend a '?' to the params string.
             false => format!("?{}", params),
         }
     }
